Rework pytest scripts to move fixtures into conftest.py
