+++ conftest.py
@@ -4,3 +4,14 @@
 # Copyright (c) 2018-2022 Linh Pham
 # stats.wwdt.me is relased under the terms of the Apache License 2.0
 """pytest conftest.py File"""
+
+import pytest
+
+from app import create_app
+
+
+@pytest.fixture
+def client():
+    app = create_app()
+    with app.test_client() as client:
+        yield client
