+++ tests/test_guests.py
@@ -6,8 +6,6 @@
 """Testing Guests Module and Blueprint Views"""
 import pytest
 
-from tests.fixture import client
-
 
 def test_index(client):
     """Testing guests.index"""
