+++ tests/test_hosts.py
@@ -6,8 +6,6 @@
 """Testing Hosts Module and Blueprint Views"""
 import pytest
 
-from tests.fixture import client
-
 
 def test_index(client):
     """Testing hosts.index"""
