+++ tests/test_shows.py
@@ -6,8 +6,6 @@
 """Testing Shows Module and Blueprint Views"""
 import pytest
 
-from tests.fixture import client
-
 
 def test_index(client):
     """Testing shows.index"""
